+++ src/main/file_open.ts
@@ -2,7 +2,7 @@ import { dialog, BrowserWindow } from "electron";
 import {
 	defaultBrowserWindow,
 	restoreWindowAndFocus,
-} from "../common/browser_window";
+} from "../common/browser_window.js";
 
 export const EVENT_FILE_OPEN_ADD_FILES = "file-open-add-files";