+++ src/main/menu_dock.ts
@@ -1,5 +1,5 @@
 import { MenuItemConstructorOptions } from "electron";
-import { fileMenuOpenItem } from "./menu_file_open";
+import { fileMenuOpenItem } from "./menu_file_open.js";
 
 export function dockMenuTemplate(): MenuItemConstructorOptions[] {
 	return [fileMenuOpenItem()];