+++ src/main/window_setup.ts
@@ -1,8 +1,8 @@
-import { BrowserWindow, app } from "electron";
+import { BrowserWindow, app, BrowserWindowConstructorOptions } from "electron";
 import url from "url";
 import path from "path";
-import { isMac, isWindows } from "../common/platform";
-import { iconPath, resourcesPath } from "../common/resources";
+import { isMac, isWindows } from "../common/platform.js";
+import { iconPath } from "../common/resources.js";
 
 const DEFAULT_WINDOW_WIDTH = 580;
 const DEFAULT_WINDOW_HEIGHT = 312;
@@ -40,7 +40,7 @@ function windowsStopFlashingFrameOnFocus(browserWindow: BrowserWindow) {
 function urlForLoad() {
 	// TODO: replace the deprecated url.format in favor of the WHATWG URL API
 	return url.format({
-		pathname: path.join(resourcesPath(), "index.html"),
+		pathname: path.join(__dirname, "..", "..", "..", "index.html"),
 		protocol: "file",
 		slashes: true,
 	});
@@ -53,7 +53,7 @@ function mainWindowLoadUrl(browserWindow: BrowserWindow) {
 const WINDOW_BACKGROUND_COLOR = "#F5F6F8";
 
 export function createMainWindow(): BrowserWindow {
-	let options = {
+	let options: BrowserWindowConstructorOptions = {
 		title: app.name,
 		show: false,
 		width: DEFAULT_WINDOW_WIDTH,
@@ -61,10 +61,14 @@ export function createMainWindow(): BrowserWindow {
 		minWidth: DEFAULT_WINDOW_WIDTH,
 		minHeight: DEFAULT_WINDOW_HEIGHT + 25,
 		webPreferences: {
-			nodeIntegration: true,
+			// TODO: disable this
+			// nodeIntegration: false,
 			// TODO: need to get this working with "true" to upgrade to Electron 12,
 			// but electron-webpack depends on it being "false" and it's been abandonded
-			contextIsolation: true,
+			// contextIsolation: true,
+			// contextIsolation: true,
+			// preload: path.join(__dirname, "..", "renderer", "index.js"),
+			preload: path.join(__dirname, "..", "renderer_main_world", "preload.js"),
 		},
 		//set specific background color eliminate white flicker on content load
 		backgroundColor: WINDOW_BACKGROUND_COLOR,