+++ src/renderer_isolated_world/i18n.ts
@@ -1,6 +1,6 @@
 import { ipcRenderer } from "electron";
-import { preloadI18nStrings, i18n } from "../common/i18n";
-import { IPC_EVENT_NAME_GET_LOCALE } from "../main/i18n";
+import { preloadI18nStrings, i18n } from "../common/i18n.js";
+import { IPC_EVENT_NAME_GET_LOCALE } from "../main/i18n.js";
 
 const ATTRIBUTE_I18N = "i18n";
 
@@ -9,7 +9,8 @@ export function setupI18n(): void {
 	translateHtml();
 }
 
-async function translateHtml() {
+async function translateHtml(): Promise<void> {
+	console.log("translateHtml");
 	const locale = await getLocale();
 	const elements = document.querySelectorAll(`[${ATTRIBUTE_I18N}]`);