+++ src/renderer_isolated_world/menu_select_files.ts
@@ -1,6 +1,6 @@
 import { ipcRenderer } from "electron";
-import { EVENT_FILE_OPEN_ADD_FILES } from "../main/file_open";
-import { selectFiles } from "./select_files";
+import { EVENT_FILE_OPEN_ADD_FILES } from "../main/file_open.js";
+import { selectFiles } from "./select_files.js";
 
 export function setupSelectFilesMenu(): void {
 	ipcRenderer.on(EVENT_FILE_OPEN_ADD_FILES, (_event, filePaths) => {