+++ src/renderer_isolated_world/table_update_row.ts
@@ -1,4 +1,4 @@
-import { sanitizeHTML } from "./sanitize";
+import { sanitizeHTML } from "./sanitize.js";
 
 export function updateRowWithExif(
 	tdNode: HTMLTableDataCellElement,